Fix PHP 8.1 deprecations (#923)
* Address return type deprecations for ArrayAccess.

* Explicitly handle float-to-int conversion.

* Address PHP 8.1 return types on IteratorAggregate.

* Address null-to-string passing deprecation for DOM XML.

* Fix more float-to-int conversion issues.

* Fix return type issues on Iterator implementations.

* Fix more not-nullable-internal issues.

* Fix PHP silliness around typing.

+++ lib/Color/Gradient.php
@@ -69,8 +69,8 @@ use function array_fill;
             $gradient[$i] = range($cStart, $cEnd, $step);
         }
 
-        $colors = array_merge($this->colors, array_map(function (int $r, int $g, int $b) {
-            return DeprecatedColor::fromRgb($r, $g, $b);
+        $colors = array_merge($this->colors, array_map(function (float $r, float $g, float $b) {
+            return DeprecatedColor::fromRgb((int)$r, (int)$g, (int)$b);
         }, ...$gradient));
 
         // remove the start color as it's already present
